Address quality findings #909-#912
Four findings from the /rhiza:quality scorecard, in one branch.

#909 deptry: add api/ to DEPTRY_FOLDERS so the FastAPI service is scanned,
and drop fastapi from the DEP002 ignore list now that its import is observed
directly. kaleido, uvicorn and python-multipart stay ignored — none of the
three is ever imported (image-export backend, CLI-invoked server, and a
Starlette-internal parser) — but each now carries a comment saying why.

#910 property tests: eight new invariants over the metric surface, chosen to
cover what snapshots cannot — ordering (worst <= best, CVaR <= VaR,
avg_drawdown >= max_drawdown), bounded ratios (exposure, monthly_win_rate),
the self-capture identity, agreement between summary() and the metric methods
it tabulates, and the truncate/smoothed_holding no-ops.

#911 complexity: truncate() drops from CC 10 to 2 by extracting the date-mask
and row-bounds helpers, with _rebuild() and _numeric_assets removing the
three-way duplication across the transforms. _stats/_reporting.py (592 lines)
splits into _reporting/_capture/_summary; _plots/_portfolio.py (588) becomes a
package of _nav/_rolling/_diagnostics mixins, mirroring _plots/_data/. Also
brings yearly_returns and rolling_beta down from CC 10.

Extracting _beta_assets exposed a line coverage had never reached: the
fallback side of a conditional *expression* is invisible to branch coverage,
so the 100% gate had been passing over it. Covered by a new test rather than
folded back into a ternary.

#912 stability docs: the pre-1.0 policy was present but buried below the
guarantee table and thin on specifics. Adds an up-front status callout and a
"Before v1.0.0" section covering what is already settled, what may still move,
how changes are announced, and pinning advice. Also corrects the stale
private-module list (_stats.py etc. are packages now; _portfolio_data.py is
gone).

Coverage stays at 100% (line and branch), both import-linter contracts hold,
and all eight gates pass.
